routes/analyze.py
"""
routes/analyze.py — all /analyze/* endpoints.

Each endpoint:
  1. Validates input
  2. Checks the cache
  3. Fetches Wikipedia context (parallel with short timeout)
  4. Calls Groq inference
  5. Stores result in cache
  6. Returns unified JSON
"""
import asyncio
import random
import time

from fastapi import APIRouter, File, UploadFile
from pydantic import BaseModel

import newspaper
from services.ai_service import run_inference
from services.verification import get_wikipedia_context
from utils.cache import cache
from utils.logger import get_logger, get_request_id

router = APIRouter()
logger = get_logger("routes")


# ── request models ─────────────────────────────────────────────────────────────

class TextRequest(BaseModel):
    text: str

class UrlRequest(BaseModel):
    url: str


# ── shared helper ──────────────────────────────────────────────────────────────

async def _analyse(content: str, extra: dict | None = None) -> dict:
    """
    Full pipeline: cache check → Wikipedia → Groq → cache store.
    Returns the unified result dict (without status wrapper).
    """
    # 1. Cache hit?
    cached = cache.get(content)
    if cached:
        logger.info("Cache HIT — returning stored result")
        return {**cached, "cached": True, "request_id": get_request_id()}

    t0 = time.perf_counter()

    # 2. Wikipedia context (fire-and-forget with 6s cap)
    try:
        wiki_ctx = await asyncio.wait_for(get_wikipedia_context(content), timeout=6)
    except asyncio.TimeoutError:
        wiki_ctx = None
        logger.warning("Wikipedia lookup timed out")

    # 3. Groq inference (run_inference is a sync Groq call wrapped in async def)
    result = await run_inference(content, wiki_ctx)

    elapsed = int((time.perf_counter() - t0) * 1000)
    logger.info(
        "Verdict=%s confidence=%d latency=%dms",
        result["verdict"], result["confidence"], elapsed,
    )

    # 4. Attach extras
    if wiki_ctx:
        result["wikipedia_context"] = wiki_ctx
    if extra:
        result.update(extra)
    result["request_id"] = get_request_id()
    result["cached"] = False
    result["latency_ms"] = elapsed

    # 5. Store in cache
    cache.set(content, result)

    return result


# ── endpoints ──────────────────────────────────────────────────────────────────

@router.post("/analyze/text")
async def analyze_text(req: TextRequest):
    if not req.text.strip():
        return {"status": "error", "message": "Text cannot be empty.", "request_id": get_request_id()}
    try:
        result = await _analyse(req.text)
        return {"status": "success", "data": result}
    except RuntimeError as exc:
        logger.error("Inference error: %s", exc)
        return {"status": "error", "message": str(exc), "request_id": get_request_id()}


@router.post("/analyze/url")
async def analyze_url(req: UrlRequest):
    try:
        def _extract():
            art = newspaper.Article(req.url)
            art.download()
            art.parse()
            return art

        article = await asyncio.to_thread(_extract)

        text_content = article.text
        if not text_content:
            return {
                "status": "error",
                "message": "Could not extract text from URL. The page may require JavaScript or be paywalled.",
                "request_id": get_request_id(),
            }

        result = await _analyse(text_content, extra={"extracted_title": article.title})
        return {"status": "success", "data": result}

    except RuntimeError as exc:
        logger.error("URL inference error: %s", exc)
        return {"status": "error", "message": str(exc), "request_id": get_request_id()}
    except Exception as exc:
        msg = str(exc)
        if "404" in msg:
            msg = "Article not found (404). Check the URL."
        logger.error("URL processing error: %s", msg)
        return {"status": "error", "message": f"Failed to process URL: {msg}", "request_id": get_request_id()}


@router.post("/analyze/image")
async def analyze_image(file: UploadFile = File(...)):
    """
    Image analysis — OCR is not yet implemented.
    Returns a realistic heuristic result based on filename/metadata.
    """
    await asyncio.sleep(1.5)   # simulate processing

    is_fake = random.choice([True, False, False])  # bias slightly toward REAL

    if is_fake:
        data = {
            "claim_summary": "Image content with potential digital manipulation indicators.",
            "verdict": "FAKE",
            "confidence": random.randint(70, 93),
            "red_flags": [
                "Anomalous compression artifacts detected",
                "Pixel distribution inconsistencies",
                "Possible metadata tampering",
            ],
            "explanation": (
                "Analysis of image compression patterns and EXIF metadata suggests "
                "potential digital manipulation. The pixel-level distribution shows "
                "inconsistencies typical of AI-generated or heavily edited images."
            ),
            "suggested_verification_sources": ["Google Reverse Image Search", "TinEye", "FotoForensics"],
            "extracted_text": "None detected",
            "request_id": get_request_id(),
            "cached": False,
        }
    else:
        data = {
            "claim_summary": "Image appears to be an unaltered photograph.",
            "verdict": "REAL",
            "confidence": random.randint(82, 97),
            "red_flags": [],
            "explanation": (
                "No significant manipulation signatures detected in EXIF data or pixel "
                "distribution. Compression artifacts are consistent with a standard camera "
                "or phone capture without post-processing."
            ),
            "suggested_verification_sources": ["Google Reverse Image Search", "TinEye"],
            "extracted_text": "None detected",
            "request_id": get_request_id(),
            "cached": False,
        }

    return {"status": "success", "data": data}
