requirements.txt
# SentinelLens — Backend Dependencies
# Install with: pip install -r requirements.txt

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── AI — Groq Cloud (OpenAI-compatible) ──────────────────────────────────────
groq>=0.8.0

# ── Article extraction ────────────────────────────────────────────────────────
newspaper3k>=0.2.8
lxml[html_clean]>=5.0.0

# ── HTTP client (async — for Wikipedia) ──────────────────────────────────────
httpx>=0.27.0

# ── Config / env ──────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── File uploads ──────────────────────────────────────────────────────────────
python-multipart>=0.0.9
